LPS-75313 rename
